+++ attributes/default.rb
@@ -1,2 +1,3 @@
 default['opsline-chef-client']['cron'] = true
 default['opsline-chef-client']['runs_per_hour'] = 2
+default['opsline-chef-client']['unregister_at_shutdown'] = false